templates: lxc-opensuse.in: Ensure cache is fully populated
We build our LXC rootfs based on a minimal openSUSE cache. However,
since we are only using this cache to bootstrap the rootfs, we need to
make sure that all packages are present there otherwise we see failures
like the following one

expanding package dependencies...
expansion error
  nothing provides net-tools
  nothing provides iproute2
Failed to download 'opensuse base'

Most of the needed packages are being pulled in indirectly as
dependencies of other packages but we should be more explicit since
dependencies could change in the future. Finally, we pass
'--no-recommends' in order to have better control of what packages are
being pulled into the cache.

Signed-off-by: Markos Chandras <mchandras@suse.de>