lxc-debian: fix getopt parsing of --mirror
When the --keyring option was added, the `,` between `mirror:` and
`keyring:` was omitted in the long option string passed to `getopt`.
This causes `--mirror` to be parsed as `--mirror:keyring`, matching
`*) break ;;` and prematurely terminating option parsing.  To avoid
this, add the missing `,`.  Also order `keyring:` before `mirror:` to
preserve lexical ordering of long options.

Fixes: eebcd76feb47 ("lxc-debian: allow using a custom keyring")
Signed-off-by: Kevin Locke <kevin@kevinlocke.name>