Fix alpine template
On a Debian sid system with lxc 5.0.3 and the latest snapshot of lxc-templates,
attempting to create an alpine container fails with the error "mknod:
dev/console: File exists". This had been reported a while back in Debian
bug #998095 and partially addressed in #46.

I performed some testing of the resulting alpine container, and things appear
to be working correctly.

Signed-off-by: Mathias Gibbens <gibmat@debian.org>